feat: refactor package structure of punk use-cases
