+++ app/src/main/java/com/example/meditationcomposeapp/model/usecase/punk/network/GetBeerByIdUseCase.kt
@@ -1,4 +1,4 @@
-package com.example.meditationcomposeapp.model.usecase.punk
+package com.example.meditationcomposeapp.model.usecase.punk.network
 
 import com.example.meditationcomposeapp.data_source.repository.punk.PunkRepository
 import com.example.meditationcomposeapp.model.entity.NetworkResponse
@@ -6,15 +6,15 @@ import com.example.meditationcomposeapp.model.entity.beer.Beer
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
-interface GetRandomBeerUseCase {
+interface GetBeerByIdUseCase {
     suspend operator fun invoke(beerId: Int): Flow<NetworkResponse<Beer>>
 }
 
-class GetRandomBeerUseCaseImpl @Inject constructor(
+class GetBeerByIdUseCaseImpl @Inject constructor(
     private val punkRepository: PunkRepository
-): GetRandomBeerUseCase {
+) : GetBeerByIdUseCase {
 
     override suspend fun invoke(beerId: Int): Flow<NetworkResponse<Beer>> =
-        punkRepository.getRandomBeer()
+        punkRepository.getBeerById(beerId)
 
 }
